+++ com/oauth2/authorize/http/response/types/codeidtokentoken.js
@@ -70,7 +70,7 @@ exports = module.exports = function(idts, ats, acs, logger, C) {
     });
 };
 
-exports['@implements'] = 'http://i.authnomicon.org/oauth2/authorization/http/ResponseType';
+exports['@implements'] = 'module:oauth2orize.RequestProcessor';
 exports['@type'] = 'code id_token token';
 exports['@require'] = [
   '../../../../../sts/id',