+++ com/oauth2/authorize/http/response/types/codetoken.js
@@ -58,7 +58,7 @@ exports = module.exports = function(ats, acs, logger, C) {
     });
 };
 
-exports['@implements'] = 'http://i.authnomicon.org/oauth2/authorization/http/ResponseType';
+exports['@implements'] = 'module:oauth2orize.RequestProcessor';
 exports['@type'] = 'code token';
 exports['@require'] = [
   'http://i.authnomicon.org/oauth2/AccessTokenService',