+++ com/oauth2/authorize/http/response/types/idtoken.js
@@ -41,7 +41,7 @@ exports = module.exports = function(idts, logger, C) {
     });
 };
 
-exports['@implements'] = 'http://i.authnomicon.org/oauth2/authorization/http/ResponseType';
+exports['@implements'] = 'module:oauth2orize.RequestProcessor';
 exports['@type'] = 'id_token';
 exports['@require'] = [
   '../../../../../sts/id',