+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseType`
+interface to `module:oauth2orize.RequestProcessor`.
 
 ## [0.0.4] - 2021-12-15
 ### Added